+++ src/Mvc/Model/KytePage.php
@@ -136,6 +136,15 @@ $KytePage = [
 			'date'		=> false,
 		],
 
+		'is_js_module' => [
+			'type'		=> 'i',
+			'required'	=> false,
+			'size'		=> 1,
+			'unsigned'	=> true,
+			'default'	=> 0,
+			'date'		=> false,
+		],
+
 		'use_container' => [
 			'type'		=> 'i',
 			'required'	=> false,