+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Refactor code and remove unreachable statements following a throw.
 * Add member methods for deleting or purging retrieved objects.
+* Add ability to mark JS code in a page as a module.
 
 ## 3.6.4